move trpc sw strategy to a single file
